normalized bundle for new format and separated argparse from logging config for sake of testing
